git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@13841 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/dds.c
@@ -1018,7 +1018,7 @@ static size_t CompressAlpha(const size_t min, const size_t max,
   codes[6] = 0;
   codes[7] = 255;
 
-  for (i=1; i < steps; i++)
+  for (i=1; i < (ssize_t) steps; i++)
     codes[i+1] = (unsigned char) (((steps-i)*min + i*max) / steps);
 
   error = 0;
@@ -1307,7 +1307,7 @@ static void CompressRangeFit(const size_t count,
       VectorCopy43(points[0],end);
 
       min = max = Dot(points[0],principle);
-      for (i=1; i < count; i++)
+      for (i=1; i < (ssize_t) count; i++)
       {
         val = Dot(points[i],principle);
         if (val < min)
@@ -1342,7 +1342,7 @@ static void CompressRangeFit(const size_t count,
   codes[3].y = (start->y * (1.0f/3.0f)) + (end->y * (2.0f/3.0f));
   codes[3].z = (start->z * (1.0f/3.0f)) + (end->z * (2.0f/3.0f));
 
-  for (i=0; i < count; i++)
+  for (i=0; i < (ssize_t) count; i++)
   {
     bestDist = 1e+37f;
     bestj = 0;
@@ -1547,13 +1547,13 @@ static MagickBooleanType ConstructOrdering(const size_t count,
 
   o = order + (16*iteration);
 
-  for (i=0; i < count; i++)
+  for (i=0; i < (ssize_t) count; i++)
   {
     dps[i] = Dot(points[i],axis);
     o[i] = (unsigned char)i;
   }
 
-  for (i=0; i < count; i++)
+  for (i=0; i < (ssize_t) count; i++)
   {
     for (j=i; j > 0 && dps[j] < dps[j - 1]; j--)
     {
@@ -1567,7 +1567,7 @@ static MagickBooleanType ConstructOrdering(const size_t count,
     }
   }
 
-  for (i=0; i < iteration; i++)
+  for (i=0; i < (ssize_t) iteration; i++)
   {
     MagickBooleanType
       same;
@@ -1593,7 +1593,7 @@ static MagickBooleanType ConstructOrdering(const size_t count,
   xSumwSum->z = 0;
   xSumwSum->w = 0;
 
-  for (i=0; i < count; i++)
+  for (i=0; i < (ssize_t) count; i++)
   {
     DDSVector4
       v;
@@ -2780,9 +2780,9 @@ static void WriteFourCC(Image *image, const size_t compression,
         alphas[i] = -1;
       }
 
-      for (by=0; by < rows; by++)
+      for (by=0; by < (ssize_t) rows; by++)
       {
-        for (bx=0; bx < columns; bx++)
+        for (bx=0; bx < (ssize_t) columns; bx++)
         {
           if (compression == FOURCC_DXT5)
             alpha = ScaleQuantumToChar(GetPixelAlpha(p));
@@ -2798,7 +2798,7 @@ static void WriteFourCC(Image *image, const size_t compression,
           p++;
 
           match = MagickFalse;
-          for (i=0; i < count; i++)
+          for (i=0; i < (ssize_t) count; i++)
           {
             if ((points[i].x == point.x) &&
                 (points[i].y == point.y) &&
@@ -2836,7 +2836,7 @@ static void WriteFourCC(Image *image, const size_t compression,
           }
         }
 
-      for (i=0; i < count; i++)
+      for (i=0; i < (ssize_t) count; i++)
         points[i].w = sqrt(points[i].w);
 
       if (compression == FOURCC_DXT5)
@@ -2945,7 +2945,7 @@ static MagickBooleanType WriteMipmaps(Image *image, const size_t pixelFormat,
   columns = image->columns;
   rows = image->rows;
 
-  for (i=0; i<mipmaps; i++)
+  for (i=0; i< (ssize_t) mipmaps; i++)
   {
     resize_image = ResizeImage(image,columns/2,rows/2,TriangleFilter,1.0,
       exception);
